fix: Apply limit after querying
Change the implementation of inferenceql.query.lang.eval/eval for SELECT
nodes such that (1) no longer applies the limit before executing the
query, and (2) executes the query on batches of the source dataset
rather than the entire input sequence. This will protect against looping
infinitely when working with an infinite sequence that does not have
values that satisfy a WHERE clause.

Change inferenceql.query.lang.select/remove-placeholders-xform such that
it preserves metadata.

Add a test that verifies that WHERE and LIMIT interact correctly.